Makefile: Simplify letting GNU make implicit rule link in a C++ way
It turns out that the GNU make implicit rule for "%: %.o"
and for "%: %.cc" are identical except for the used $(LINK.o)
or $(LINK.cc), respectively. So it's possible to have the same effect
of fully spelling out a new linking rule by just overriding the
link-from-object-files linker command variable!

Thanks, @Isilkor!