Makefile: Add support for automatic dependency generation
With this, a change to, say, helloworld/helloworld.hh, will remake
helloworld/helloworld.o, helloworld/main.o, and helloworld/helloworld --
although the .hh header file is not specified manually in the Makefile.

This lets the build process require bash, for the complicated recipe
to generate the automatic dependency information, and post-process it.
This is accomplished without the use of a temporary intermediate file.

Also:
* Ignore the auto-generated dependency files from version control.
* Remove them on "make clean".
* Include them from the main Makefile. This will give spurious messages
  that should not be fatal, as the required files can be made from
  existing files and make will restart, then:

    Makefile:35: simple.deps: No such file or directory
    Makefile:35: helloworld/helloworld.deps: No such file or directory
    Makefile:35: helloworld/main.deps: No such file or directory