Target C++14
I'm going to use std::quoted() and that appears to require C++14.

Also Debian stable (Debian 8) g++ seems to support C++14 just fine,
if told to use it (via -std=c++14), and Debian testing
(upcoming Debian 9) even defaults to it (or rather, to the GNU dialect
of it).

So, let's just use it.