[ls-gui-cvn/lsgui] Indicate error number + total error count: Rewrite error accumulation
With this, on unsuccessful scroll-to-bottom, the user may look at the
error message number and notice that it's, e.g., "4/5" so he'll have to
scroll manually an additional line.

To implement, we don't keep previous markup string, but a list
(std::vector) of previous error messages. Then on each change,
the new total error count will be inserted at the right place.

This kills the fall-back to set_text(), though. If the markup escape
should fail, the message just gets displayed as N.A. (not available)...