[ls-gui-cvn/lsgui] Add menubar loading code
The menubar is there, now, and the proper signal-connected
action handlers are run.

The test action that was missing gives a greyed-out menu item.
This properly reflects the fact that it can't be run as the action
doesn't exist.

The accelerators were not expected to work, and are greyed out to match.

So, everything seems to be quite good at the moment.