[ls-gui-cvn/lsgui] Implement location history
* Make LsGui::set_location_str(ustring) the new set_location_str(),
  which fetches the current history element as new location_str_.
* Add a new set_location_str(ustring) which manages the history
  and then calls set_location_str().
* Let reload just call set_location_str() without arguments, so,
  stay with the same history and at the same history position.
* Implement backwards and forward movement in history, by advancing
  the location history position iterator.
* Have several fail-safes to just ring the error bell and stay
  with the current state in case of problems.

This all seems to work quite well, at the moment.

Fun fact: The implementation of forward will be able to advance
too far, so that the last forward and subsequent reloads will complain
about being outside of history. But it doesn't crash, just complain
to stderr and ring the error bell! \o/