[ls-gui-cvn/lsgui] Refactor: Add history_is_valid() and friends
This encapsulates the checks used by various parts of the code.
The intention is to later disable an action temporarily (so that
it'll be greyed-out in the UI elements, like menubar & toolbar,)
based on the same check that would be used by itself to know
whether it should fail early.