{CodeLite} Bugfix: Prevent overwriting Makefile
Enable custom build for project misc. (Also disable the project.)

Before, it created a misc.mk and a Debug/ subdir, which would both
be reasonable; but also has overwritten the Makefile with a
CodeLite-generated one. Apparently, if only a single project
does *not* have custom build set, CodeLite will always overwrite
the Makefile... The fact that it's been made part of the misc project
does not seem to change anything in this regard.