[ls-gui-cvn/lsgui] Location completion: Esc 2x to reset entry to previous location
The idea is to have one press of Escape key to close the completion
popup window, and another press to reset the entry to the previous
location string.

It's possible there is still a bug in the logic, but this is better now
than losing the full input string when just trying to close the popup.

* Also update TODO.