[ls-gui-cvn] Refactor: Rename Dirent::get_name() to get_ent_name()
This is for clarity; the Dirent class is often instantiated as a variable
with name dir. Then, the old call would have been dir.get_name(), which
would suggest that it's the name of the directory, not of the entry.
After this, it'll be dir.get_ent_name() which tells what it really is.