[ls-gui-cvn/util] cvn::fs::readlinkat(): Use byte size, not element count
Both are equal, in this case, but it was semantically wrong.

If anything, I would have needed to convert the "ret" return value
from byte size to element count, too, before accessing the "buf" array.
But as it's a char array, indexing is in chars, so in bytes, so
it's the same anyhow.