[ls-gui-cvn/stat-cvn] get_{user,group}(): Throw std::system_error exceptions on detectable error
Experimentation suggests that, at least on glibc on GNU/Linux
(upcoming Debian 9), a "not found" is not considered an error
and, therefore, leaves errno at the previously set value.
(In this case, 0.)

But, with this we should be prepared to receive and maybe handle
some weird error when using advanced NSS (Name Service Switch)
configurations.