[ls-gui-cvn/lsgui] Bugfix: Avoid accidental copying of Gtk::FileChooserDialog
With newer C++ implementations, the copy seems to have been
implicitly left out, so it worked. On Debian 8 'jessie' / g++ 4.9.2,
it actually attempted compiling a copy operation, which is
explicitly made impossible by the Gtk::FileChooserDialog declaration.

So just use proper initialization instead of assignment to fix
this part of the loads of build errors on Debian stable.