[ls-gui-cvn/dirent-cvn] Add get_ent_type() and supporting enum class EntType
With this, the type of a directory entry may be seen
without the need for doing an additional lstat().

Types are, e.g., regular file, directory, symlink, ...

If the struct dirent's d_type field should not be available,
this will give a compile-time warning and always return
the unknown type.