[ls-gui-cvn/lsgui] When position of last slash changes, update location completion
That also means to update the completion when the user
backspaces over a slash (as in "directory separator").

* Also update TODO.

Previously, the location entry completion was updated
when descending into a subdirectory by typing a slash '/'.
Now it'll be updated also when returning to a parent
directory by removing the last slash.