[build system] Use describe-cvn to obtain ls-gui-cvn version under git
This should make it possible to, say, develop another project
in this same repository and let the ls-gui-cvn version stay
the same. (Ideally, directly at its tagged release version.)

This does not help with outside-of-git builds, though. I suggest
getting an archive from the release-tagged version directly and
build from that.