[build system] Bugfix: Makefile: Set LDLIBS, not LDFLAGS
This gives the correct ordering of previously compiled .o files
and the -l flags.

It seems that the GCC on Travis CI, g++ 4.8.4, still needs correct
ordering, and the later GCCs (g++ 4.9.2 in Debian 8 stable) don't.

So, this shouldn't have an effect on Debian stable onwards, but
will hopefully fix the build on Travis CI / Ubuntu trusty 14.04.