[build system] Makefile: On dependency generation, silence compiler error messages
Avoid causing confusion about a "missing" config.h,
with previously red error messages from the compiler
that a header file would be missing. It'll be generated later,
so silence the compiler's stderr. If it has something to say,
it can well say it during normal compilation of the same source.
Plus, if it exits with a non-zero exit code, the shell snippet
will do its own output to stderr.