[build system] Makefile: Add targets deps, cleanconfig, cleandeps
It might be desirable to completely get rid of the dependency makefiles
without also losing all the build results. Even make -k (keep going)
does not cope with a header disappearing from the source tree. So,
use "make cleandeps" in such a situation.

Also, it could be desirable to just clean the config header(s). So
don't do that as an additional step in make realclean, but have a
separate target for that and have a new realclean depend on that.