[CI] IRC notifications: Did not seem to work
It built fine on Travis CI, but on IRC nothing arrived.

Is it configured in a wrong way? Or maybe are there general problems
currently at Travis CI or freenode? Leaving the config unchanged
(except for a comment), for now; but leaving this message so as to
not give the impression that it would have worked.