[CI/osx, ls-gui-cvn/time-cvn] Insert additional duration_cast into cvn::system_time::from_timespec()
On other systems, e.g., macOS/osx/Apple, the system clock may have
less precision than nanoseconds... So an explicit cast is needed,
there, and may lose the filesystem-provided nanoseconds precision!

TODO: Let this depend on a build-system test instead of on the platform?

TODO: Make the duration_cast unnecessary by storing time
in a non-system_clock, explicitly nanoseconds-based
representation... (Or is this not worth the effort as
maybe no-one will ever need the 3 digits missing on macOS?)

[CI/osx] This is expected to (at least partially) fix the macOS build